refector: separating the folder structure (all)
