+++ frontend/src/components/topbar/Activity/ActivityCloseButton.js
