+++ frontend/src/containers/main/column/card/CardContainer.js
@@ -1,0 +1,35 @@
+import React, { useState } from 'react';
+import axios from "axios";
+import Card from "../../../../components/main/column/card/Card";
+
+const CardContainer = ({ title, body, index, setCardList, columnId, cardId, previousCardId }) => {
+    const [isCardDelHover, setIsCardDelHover] = useState(false);
+
+    const onClickDeleteCardHandler = () => {
+        setCardList((cardList) => cardList.filter((_, i) => i !== index))
+        axios.delete(`api/columns/${columnId}/cards/${cardId}`)
+    };
+
+    const onDoubleClickEditCardHandler = () => {
+        setCardList((cardList) => {
+            const left = cardList.slice(0, index);
+            const right = cardList.slice(index + 1);
+            return left.concat({ title, body, isInput: true, previousCardId, cardId, columnId }, right);
+        });
+    };
+
+    const onMouseEnterHandler = () => setIsCardDelHover(true);
+    const onMouseLeaveHandler = () => setIsCardDelHover(false);
+
+    const data = { title, body };
+    const onEvents = {
+        onClickDeleteCardHandler,
+        onDoubleClickEditCardHandler,
+        onMouseEnterHandler,
+        onMouseLeaveHandler,
+    };
+
+    return (<Card data={data} isCardDelHover={isCardDelHover} onEvents={onEvents} />)
+};
+
+export default CardContainer;
